[script][game] Implement GetAttemptedSpellTarget
The patch adds GetAttemptedSpellTarget to get the last target of a
spell, as well as infrastructure to support other GetAttempted*
routines.

Every object now stores a history of its combat actions. For now, only
spell actions are added, but it should support attack, move, and other
actions as well.

There is an existing Character::getAttemptedAttackTarget() that is
used for GetAttemptedAttackTarget routine, but I think it doesn't
follow semantic of GetAttemptedAttackTarget. It only checks for the
current action, while the documentation suggests that it should search
for previous actions if the current one has a different type. For
example, when an attack is followed by a move,
GetAttemptedAttackTarget should return the attack even when it is no
longer "current".

There is some overlap between Combat::attackHistory() and
Creature::combatActionHistory():

  - Combat::attackHistory() contains only "attacks", but they are
    augumented with resultType (hit, miss) and damage.

  - Creature::combatActionHistory() contains all actions performed by
    creature during combat. This includes move, spells, and attacks
    as well.