[script] Implement GetLastAttackAction
The function used to be called GetLastAttackObjectAction, but the name was
inconsistent with other KOTOR tools.

The function should return a type of the last attack action (AttackObject,
CastSpell, etc.)

For this routine (as well as for other GetLast* routines) the patch adds a
combat history that keeps all past attacks for each creature.