[game][script] Implement GetCreatureTalentBest
For GetCreatureTalentBest and other talent-related routines, it is
benefitial to store feats and spells in a sorted data structure with
random access.

This allows us to find a feat/spell by category (binary search), and
then iterate over elements to find the one matching criteria.

Same approach should work for GetCreatureTalentRandom, but with
std::lower_bound + std::upper_bound.

It should also be benefitial to replace std::unordered_map with
another array (sorted by type). With low number of elements (~100) it
should be better for performance, but this is off-topic for this
patch.

The patch also initializes spells and feats globals.