[nfc][script] Print a unique value number for each operand and result
The patch adds an id to each variable, and makes all variables
immutable. This is not enforced, but instructions do not change values
in place now.

Script3 log is now more verbose as it prints all operands and results
for each instruction:

    DEBUG [main][script] Instruction: 0000048f CONSTI 39 () -> (%106:39)
    ...
    DEBUG [main][script] Instruction: 00001fef CONSTO 0 () -> (%399:4)
    ...
    DEBUG [main][script] Action: 1ff5 GetLocalBoolean(%399:4, %106:39) -> %400:0
    ...
    DEBUG [main][script] Instruction: 00002010 CONSTI 0 () -> (%401:0)
    DEBUG [main][script] Instruction: 00002016 GTII (%401:0, %400:0) -> (%402:0)
    DEBUG [main][script] Instruction: 00002018 JZ 00002044(44) (%402:0) -> jump

Here, the caller object %399:4 is passed to GetLocalBoolean with a
constant %106:39. GetLocalBoolean returns %400:0, which is compared
against constant %401:0 to produce %402:0. JZ takes this value and,
because it is zero, jumps forward to address 00002044.