[script] Use Object flags to implement GetLast* routines
These are really tied to what ai scripts expect, and are supposed to
be used only in the respective callbacks (except for
GetLastHostileActor, which seems to be used elsewhere as well).

Its possible to implement them around a generic action history, but if
we cannot change their semantic, then what's the point.