[game] Implement GetLastAttackAction and GetLastAttackResult
These routines are used in DetermineCombatRound subroutine of
k_ai_master, which is called from almost all combat-related
callbacks. This makes it difficult to implement GetLastAttackAction
and GetLastAttackResult as arguments like other GetLast* routines.

The patch adds a state to Combat to keep the last attack for each
attacker instead of immediately releasing it in finishRound.

An attack may not have an associated action if it was added by
CutsceneAttack routine.