[game] Complete PlayAnimation actions immediately for loop anims
Loop animations never finish, so if we keep the corresponding
PlayAnimation around, it is going to completely stall action queue of
the target object.

This happens on Endar Spire after the jedi fight. Reinforcement units
have the following actions:

  Move -> PlayAnimation(LoopingReady) -> DoCommand.

The last DoCommand is supposed to change faction to Hostile, but we
used to never reach it because PlayAnimation was looping.

The patch fixes issue #3 "Enemies after jedi fight spawn as neutral"
and unblocks the door to the bridge on Endar Spire.