Fix text rendering
The cause of the lines in the text rendering were due to the texture
wrapping being set to ClampToBorder instead of Repeat. Due to floating
point inaccuracy, when the uv lined up to the edge of the font texture,
the texture would be read beyond its edge and since it would clamp to
the border, a sharp line would be rendered. By setting the texture
wrapping to Repeat, when the texture is read past its edge, it wraps
around and reads a correct transparent pixel instead.