[script] Introduce arguments to ExecutionContext
This is the first patch to replace fields of ExectuionContext with a
list of "tagged" variables (arguments). These are used to store data
for GetLast* routines: GetLastOpenedBy, GetEnteringObject, etc. They
are local to a script run, but can be copied over to other script runs
that originate from STORE_STATE or ExecuteScript.

This approach makes it easier to add new arguments, and run scripts
with a list of arguments. This should also fix some unrelated routines
that used to share the same field of ExcecutionContext. For example,
GetLastOpenedBy and GetEnteringObject used to be functionally
identical, because they both used triggererId.