[game] Force targets into combat when they are attacked
When a target is not yet attacking back, we still need to tick its
combat routines. This way AI scripts have more opportunities to
dispatch a combat action.

We also need to activate combat: k_ai_master checks for GetIsInCombat,
and does not respond back to attacks otherwise.

This slightly improves combat for #7.