[game] Face destination after MoveToObjectAction
When an enemy is heard, but not seen by a creature, k_ai_master issues
a MoveToObjectAction in an attempt to find it.

However, MoveToObjectAction completes early when the target is already
in close range, so the creature does not move. Still, it must face the
target in order to make it "seen" (i.e. turn to bring the target into
FOV).