[graphics] Fix text rendering (#32)
The cause of the lines in the text rendering were due to the texture
wrapping being set to ClampToBorder with incorrect border color.

Due to floating point inaccuracy, when the uv lined up to the edge of
the font texture, the texture would be read beyond its edge and since
it would clamp to the border, a sharp line would be rendered. By
setting a transparent border color, when the texture is read past
its edge it takes a transparent color and blends with background.