[game] Refactor Creature damage and death functions
Damage effect used set health points directly, and cause die() to be
called on the next updateHealth. Tracking the object that actually
caused damage was difficult and we need those to implement
GetLastDamager for onDeath scripts.

The patch removes `Object::die()` function and replaces it with
`Object::damage(amount, damager)` to make an explicit link between
damage (or death) and the object that caused it.

If there is no object (e.g. when DeathEffect is applied from a
script), damager is NULL.

When amount is `std::numeric_limits<int>::max()`, we treat this as a
special case to ignore isMinOneHP (used for a DeathEffect).

Subsequent patches make use of damagerId in `Creature::runDeathScript`
and pass it as script arguments.