[game] Implement GetLastHostileActor
GetLastHostileActor can be used for arbitrary objects, so it cannot be
an argument like many other GetLast* routines. We have to keep this as
another state of Combat.

LastHostileMap keeps a relation between the target and the last actor
that attacked the target. We probably could've fished out this
relation by iterating over all rounds and last rounds as well, but
keeping a separate map is easier and good enough for now.