[game] Do not consider dead enemies as enemies for GetNearestCreature
While there is a dedicated criteria to only match "alive" creatures,
k_ai_master doesn't use it when looking for a target to attack.

Before the patch, k_ai_master used to select the same target
regardless of isDead condition, so it got stuck the moment its target
died: it couldn't attack anymore because the target is dead, and
couldn't find another target.

This change makes k_ai_master to ignore a dead target, and select a
new one.

The patch improves combat for #7.