[game] Do use FOV to determine visibility
The idea is great, but k_ai_master does not play well with it: when a
creature is not engaged in combat, it only reacts to "seen" objects
and completely ignores objects which are only "heard".

Without the patch, a creature can stand right behind enemies and they
never trigger (unless it gets into their 60 degree FOV).

The patch makes visibility perception work in all directions, but
still requires line-of-sight. If an enemy is behind a door, it is
counted as not "seen", but maybe "heard" if is within hearing range.