[game] Load rooms before loading objects
Rooms must be initialized first, because the first thing an area
checks when landing an object is the room it belongs to. In order to
find a room, it must be initialized and have its walkmeshes loaded
into scene graph for testElevation to work.

Since a room is "visible" by default, and all its walkmeshes are
enabled, all we need to do here is to ensure that all rooms loaded
before any object loads.