[game] Set object visibility in Room::addTenant
When an object moves into a room, it should take its visibility
state. This is to ensure consistency with setVisible(bool) function
which propagates new status to all objects.

There is no issue that require this fix (so far), but there is no
point in keeping an object "visible" (i.e. active) in a room where
everything else is disabled, including walkmeshes.