[game] Refactor combat and attack implementation to use actions
Attacks used to be implemented completely in Combat class: attack and
damage rolls, animation, projectiles, round scheduling, duel matching,
etc. It was quite complex even though it only supported basic
attacks.

In order to add more combat features such as feats and spells, we need
to refactor Combat and move all action-specific logic into the
corresponding actions.

The patch implements this for AttackObjectAction (basic attack) and
CutsceneAttackAction. Combat is now responsible only for scheduling
and duel-matching and have no dependency on what kind of action is
performed during a round.

This approach also simplifies how round history is maintained. There
used to be dedicated LastAttackMap and LastHostileMap to implement
script routines. Now we keep all past rounds in the round queue, so
these routines can access and filter it as necessary.

CutsceneAttackAction is a new action class to implement CutsceneAttack
routine. It is similar to a regular attack, but animation, attack roll
and damage amount are fixed and provided by the routine.

Animation selection is moved from creature.cpp to the corresponding
attack action. Feats and spells and other actions can follow this
design and not clutter creature.cpp.

Projectile handling is also improved. Multiple projectiles per attack
are supported, and they should match the animation. A single blaster
shoots 2 projectiles for a basic attack, and a rifle shoots 3 at
different time points.

The patch also improves actions in general. There is a new cancel()
function that acts as a hook to cleanup an action before it is
cancelled and removed. For now it is only used to reset projectiles.

Dead actors now cancel and clear all actions to not leave actions
hanging in incomplete state forever. This was also an issue for
projectiles, because we need to either cancel() or finish an action to
reset them.