[system] Add SmallSet container
SmallSet is a container similar to std::set, but it uses a SmallVector
to store elements co-allocated with the object. When the number of
elements exeeds capacity of co-allocated storage, SmallSet makes a
heap allocation and copies all elements there.  Key differences of
SmallSet from std::set:

  - SmallSet is not always sorted. When it uses a co-allocated storage,
    new unique elements are appended without maintaining sort order.

  - The underlying container is a SmallVector, so elements are stored
    contiguously in a single memory block.

  - Even though a SmallSet can expand to a heap allocation,
    performance of insert() and erase() operations is poor for large
    number of elements.

Note that SmallSmall is usually expensive to copy by value, so the
copy constructor is deleted. Use ISmallSet to pass it by reference, or
std::set if copying is required.

Note that not all functions of std::set are supported yet.