[system] Add ArrayRef<T> class to pass sequences as parameters
ArrayRef provides a unified interface for an array-like sequence. For
example:

  int sum(ArrayRef<int> arr) {
    for (int v : arr) {
      // ...
    }
  }

The function can now be called with a std::vector, a C-array, a
SmallVector, with an initializer list, etc.

ArrayRef keeps a begin() pointer and a size of the container it is
initialized with. It does not copy or take ownership of the data.