[game] Support casts for Object and Action hierarchies
This allows to check that an Object (or Action) has a specific type,
and cast it to this type, if necessary. For example:

  if (isa<Creature>(object)) {
    // do something only for Creature objects
  }

Once we checked that an object is of the right type, we can use an
unconditional cast to get a reference or a pointer of this type:

  if (isa<Creature>(object)) {
    Creature &c = cast<Creature>();
    c.giveXP(42);
  }

There is also a checked dyn_cast<> that works only on pointers (or
shared_ptr):

  Creature *c = dyn_cast<Creature>(&object);
  if (c) {
    c->giveXP(42);
  }

Here, dyn_cast returns a nullptr when object is not actually a
Creature. This can also be refactored as:

  if (auto *c = dyn_cast<Creature>(&object)) {
    c->giveXP(42);
  }