[game] Complete FollowAction and FollowLeaderAction
FollowAction and FollowLeaderAction used to never complete. AI scripts
routinely call ClearAllActions which forcefully removes these
actions. However, when combat is not active, these actions stay in the
front of the queue and do not let other actions to execute.

On Endar Spire this locked a DoCommandAction for Trask after clicking
on the "level-up" door on the bridge.