[game] Refactor projectiles
Projectiles used to be created and updated in Combat, but with attacks
moving out of combat we have to refactor projectiles as well.

Projectiles are created for every blaster or rifle shot triggered by
basic attacks actions, feats, cutscene attack actions, etc. A single
action may trigger multiple projectiles at certain time points that
match the animation.

Implementation for now is very basic: all projectiles follow the same
trajectory and hit the same point (impact node) of the target. Later
this can be extended to add some randomness and handle "missed"
attacks, for example.

There is a generic TimeEvents data structure to bind "events" to time
points. ProjectileSequence uses TimeEvents to determine when it is
time to trigger another projectile. Each projectile is then updated
until it reaches the target, and then removed from the scene.