[game] Parse projectiles 2da files
The patch adds Projectiles - a new Game component that keeps
information on how to fire projectiles for different creature types.

Projectiles are initialized from 2DA files at startup:

  - `weapondischarge.2da` contains timings for projectiles, as well as
  a mask for main/offhand firing, and the number of shots and misses.

  - `droiddischarge.2da` is an index to `weapondischarge.2da` based on
  appearance type for droid creatures.

  - `animations.2da` is used to find animation ids and match them to
  columns of `droiddischarge.2da`.

The original design is quite unclear. However, here is what we were
able to figure out implement:

`weapondischarge.2da` is split for humanoids and droids. For humanoids
it has 4 rows for each ranged wield type: blaster, dual blaster,
rifle, repeating blaster.

For droids it is more complicated. First we start at `animations.2da`
and find animation ids for b0a* animations. These ids are used as
columns of `droiddischarge.2da`. In `droiddischarge.2da`, rows contain
appearance ids for the corresponding creature (c_drdmkone, for
example). Rows of `droiddischarge.2da` are in the same order as droid
rows of `weapondischarge.2da`. The only "trick" here is to exclude
"empty" rows of `weapondischarge.2da`, as well as first 2 non-empty
droid rows. The latter is weird, but everything else matches. This
could be a bug in the original engine that was worked around this way.