[game] Lock actions in progress
CancelActions should not affect characters in the middle of attack
animation, for example.

The patch adds a function to lock() an action. When iterating over
actions in Object::clearAllActions, we start removing actions from the
back of the queue (most recently added actions). Actions in progress
should be at the front of the queue, so we stop stop on the first
locked action.

When objects are dead, all actions *must* be removed, because they
will not get an update anymore. Added a "force" flag to override locks
in this case.