[game] Fix damager id for runDeathScript
Scripts require either a valid id, or kObjectInvalid (1), but we used
to pass 0 (aka kObjectSelf). This used to be handled correctly for
runDamagedScript, but runDeathScript below used the parameter as is.

This patch fixes a regression introduced in 92dd2051:
"[game][nfc] Keep damagerid instead of a shared_ptr in DamageEffect" (#51)