[system] Add StringRef
StringRef is used to efficiently operate strings and substrings that
do not have to be null-terminated. Functions such as substr() or
strip() do not allocate memory, and return a new reference to the same
memory.

Similarly to ArrayRef, StringRef does not own memory it refers to, so
it is only suitable for passing a reference to the original object
between functions.