[system] Add TextBuffer
TextBuffer is a linear container with a file-like interface: seek(),
tell(), read(), write(), etc. This is used to emulate a text file with
a cursor, read and display it line-by-line via readline() and
readlineReverse().

The underlying memory may be reallocated after write() calls.
TextBuffer operates with indices (offsets) which are not affected by
reallocation.