[engine] Add an option to execute console commands at startup
The patch adds --commands-file command line option that takes a file
name:

   ./engine --commands-file path/to/commands.txt

When this option is specified, once the engine is initialized, it
reads the file and executes each line as a console command.

This is useful for debugging to quickly reach the required game
state. For example, you can warp to a particular module, set
variables, turn on debug rendering, spawn objects, etc.