[game] Add new console commands
The patch adds the following commands:

- `spawncreature creature_template [id]` spawns a creature from a
template (`c_drdmktwo`, for example) and optionally assigns it the
given `id`. Having a fixed id is useful if we need to select this
creature later using `selectobjectbyid` command. All creatures are
spawned as Neutral at the position of the party leader.

- `spawncompanion creature npcindex [id]` spawns a party member from a
template (`p_juhani`, or `c_drdmktwo`, for example) with a given
npcindex (from 1 to 9) and an optional object id.

- `selectobjectbyid id` selects an object with the provided `id`. All
objects are supposed to have unique ids. Use `info` command to find an
object id of the selected object.

- `selectleader` selects the current party leader. This can be useful
to effectively "cancel" selection of another object, so that follow-up
commands can target the party leader instead.

- `setfaction number` changes faction of the selected creature (or the
party leader). See enum class Faction in types.h for a complete list,
but the following are typical factions to set: 1 for Hostile, 2 for
Friendly, and 5 for Neutral.

- `setposition x y z` changes position of the selected creature (or
the party leader). X, Y, Z are world coordinates. The easiest way to
find coordinates in game is to position a creature there and use
`info` command.

- `professionaltools` adds different kinds of weapons and consumables
to the inventory of the selected object (or the party leader). This is
useful to test combat features (along with `spawncreature` and
`setfaction`).

- `killroom` kills all hostile creatures in the same room as the
selected object (or the party leader). Note that "hostile" is computed
relative to the selected object. Therefore, if you select an enemy,
`killroom` is going to kill all enemies of that enemy - typically all
party members.