[game] Remove restriction on the number of party members
The patch removes this restriction for game logic, but GUI still
supports only 3. The only way to switch to extra party members is to
use TAB key.

This feature is primarily needed for testing with `spawncompanion`
console command, but it can be enabled to control "friendly follower"
NPCs in the future.

The patch also restores Carth as the default companion for K1.