[game] Enable projectiles for CutsceneAttackAction
CutsceneAttack is a special action that is supposed to be used only by
scripts for 'cinematic' fights in dialogs. It allows to use any
animation ID as an attack regardless of the attacker's wield
type. This makes determining the corresponding projectiles a bit
tricky.

Given an animation ID, we first lookup the corresponding name. Let's
take b5a1 as an example:

  - 'b' is a blaster animation.
  - '5' is a wield type - a single blaster.
  - 'a' is a key for 'attack'.
  - '1' is a 'basic' attack.

With this data we can find the corresponding sequence of projectiles
required for this attack.