[game] Fix buttons on level up screen
Buttons and labels have _selectable field set to true and false
respectively. It is not required to set them explicitly. Here it
caused a problem, because all buttons were not selectable, and
therefore did not registered mouse clicks.

This patch allows to complete a level up process, and unblocks the
door on Endar Spire bridge.